requirements.txt
# Database adapter
dbt-duckdb[glue]>=1.3.1

# extra features
sqlfluff
shandy-sqlfmt[jinjafmt]
aws-cdk-lib
dbt-osmosis
mypy
black
aws-lambda-powertools